requirements.txt
accelerate==0.19.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.6.15
charset-normalizer==3.1.0
click==8.1.3
datasets==2.12.0
dill==0.3.6
docker-pycreds==0.4.0
exceptiongroup==1.1.1
filelock==3.12.0
frozenlist==1.3.3
fsspec==2023.5.0
gitdb==4.0.10
GitPython==3.1.31
huggingface-hub==0.14.1
hydra-core==1.3.2
idna==3.4
importlib-resources==5.12.0
iniconfig==2.0.0
Jinja2==3.1.2
joblib==1.2.0
lightning-utilities==0.8.0
MarkupSafe==2.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
networkx==3.1
numpy==1.24.3
omegaconf==2.3.0
packaging==23.1
pandas==2.0.1
pathtools==0.1.2
pip==22.1.2
pluggy==1.0.0
protobuf==4.23.0
psutil==5.9.5
pyarrow==12.0.0
pytest==7.3.1
python-dateutil==2.8.2
pytorch-lightning==2.0.2
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
responses==0.18.0
scikit-learn==1.2.2
scipy==1.10.1
sentry-sdk==1.23.0
setproctitle==1.3.2
setuptools==61.2.0
six==1.16.0
smmap==5.0.0
sympy==1.12
threadpoolctl==3.1.0
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
torch==2.0.1
torchmetrics==0.11.4
tqdm==4.65.0
transformers==4.29.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wandb==0.15.2
wheel==0.37.1
xxhash==3.2.0
yarl==1.9.2
zipp==3.15.0
